+++ cookbooks/aws-parallelcluster-platform/resources/install_packages/install_packages_redhat8.rb
@@ -23,7 +23,7 @@ def default_packages
   # environment-modules required by EFA, Intel MPI and ARM PL
   # Removed libssh2-devel from base_packages since is not shipped by RedHat 8 and in conflict with package libssh-0.9.6-3.el8.x86_64
   # iptables needed for IMDS setup
-  %w(vim ksh tcsh zsh openssl-devel ncurses-devel pam-devel net-tools openmotif-devel
+  packages = %w(vim ksh tcsh zsh openssl-devel ncurses-devel pam-devel net-tools openmotif-devel
      libXmu-devel hwloc-devel libdb-devel tcl-devel automake autoconf libtool
      httpd boost-devel mlocate R atlas-devel
      blas-devel libffi-devel dkms libedit-devel jq
@@ -31,4 +31,34 @@ def default_packages
      libgcrypt-devel libevent-devel glibc-static bind-utils
      iproute NetworkManager-config-routing-rules python3 python3-pip iptables libcurl-devel yum-plugin-versionlock
      coreutils moreutils curl environment-modules gcc gcc-c++ bzip2)
+
+  if aws_region.start_with?("us-iso")
+    packages -= %w(openmotif-devel hwloc-devel R blas-devel dkms libedit-devel glibc-static
+       NetworkManager-config-routing-rules yum-plugin-versionlock moreutils)
+  end
+
+  packages
+end
+
+action :install_extras do
+  if aws_region.start_with?("us-iso")
+    remote_file "epel_deps.tar.gz" do
+      source "#{node['cluster']['artifacts_s3_url']}/dependencies/epel/rhel8/x86_64/epel_deps.tar.gz"
+      mode '0644'
+      retries 3
+      retry_delay 5
+      action :create_if_missing
+    end
+
+    bash 'yum install missing deps' do
+      user 'root'
+      group 'root'
+      code <<-REQ
+      set -e
+      tar xzf epel_deps.tar.gz
+      cd epel
+      yum install -y * 2>/dev/null
+      REQ
+    end
+  end
 end